@@ -1,6 +1,6 @@
 class NewsArticle {
   final String title;
-  final String imageUrl;
+  final String? imageUrl;
   final String url;
   final String description; // Add description field
   final String publishedAt;  // Add publishedAt field
@@ -8,7 +8,7 @@ class NewsArticle {
   // Constructor with required fields and optional fields for null handling
   NewsArticle({
     required this.title,
-    required this.imageUrl,
+    this.imageUrl,
     required this.url,
     required this.description, // Required description
     required this.publishedAt,  // Required publishedAt
@@ -18,7 +18,7 @@ class NewsArticle {
   factory NewsArticle.fromJson(Map<String, dynamic> json) {
     return NewsArticle(
       title: json['title'] ?? 'No Title', // Default value for null title
-      imageUrl: json['urlToImage'] ?? 'https://via.placeholder.com/150', // Default placeholder image for null image
+      imageUrl: json['urlToImage'], // Default placeholder image for null image
       url: json['url'] ?? '', // Default empty string if no URL provided
       description: json['description'] ?? 'No description available.', // Default description
       publishedAt: json['publishedAt'] ?? DateTime.now().toIso8601String(), // Default to current date/time if null
