@@ -22,9 +22,8 @@ class PaymentController extends GetxController {
   Future initSDK() async {
     _midtrans = await MidtransSDK.init(
       config: MidtransConfig(
-        clientKey: "SB-Mid-client-wYeSigdies_2dI1d",
-        // clientKey: "SB-Mid-client-v6-15hiE40WcarUI",
-        // clientKey: "SB-Mid-client-v6-15hiE40WcarUI",
+        // clientKey: "SB-Mid-client-wYeSigdies_2dI1d",
+        clientKey: "SB-Mid-client-v6-15hiE40WcarUI",
         colorTheme: ColorTheme(
           colorPrimary: Theme.of(Get.context!).colorScheme.secondary,
           colorPrimaryDark: Theme.of(Get.context!).colorScheme.secondary,
